requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
phi-framework==0.4.0flask==3.0.2